Migrate ControllersIndex to RTL.
