+++ src/components/ButtonGroup/_button-group.scss
@@ -26,6 +26,7 @@
     @include desktop {
       margin-bottom: 0;
     }
+
     @include large {
       width: auto;
     }
